Fix problems with types... strongly typed stuff is a pain in the ass. this really should've been written in C

+++ serial-driver/core/src/ares_lora_serial/serial-driver/command_response.cpp
@@ -21,7 +21,7 @@ py::object CommandResponse::specific_ret(const AresFrame::Setting &ret) {
 }
 
 py::object CommandResponse::specific_ret(const AresFrame::Led &ret) {
-    return py::cast(ret.state);
+    return py::cast(static_cast<uint8_t>(ret.state));
 }
 
 py::object CommandResponse::specific_ret(const AresFrame::Heartbeat &ret) {
